Completed initial coding of all routes. All local tests failing because of timeouts

+++ routes/api.js
@@ -38,44 +38,56 @@ module.exports = function (app) {
   app
     .route("/api/books")
     .get(async function (req, res) {
+      try {
       let arrayOfBooks;
       await Book.find().then((arrayOfBooks) => {
         //return res.json(arrayOfBooks);
         let books=[];
-        for (const {title, id, commentcount} of arrayOfBooks) {
-          books.push({ title, id, commentcount });
+        for (const {title, _id, commentcount} of arrayOfBooks) {
+          books.push({ title, _id, commentcount });
         };
-        return res.json(books);
-      });      
+        res.json(books);
+        return;
+      }); 
+    } catch (err) {
+        res.json([]);
+    }            
     })
     .post(async function (req, res) {
       let title = req.body.title;
-      if (!title) {
+      if (!title || title == '') { 
         return res.json("missing required field title");
       }
+      //Not asked for by FCC project so maybe leave it out?
       if (await Book.exists({ title: title })) {
-        return res.json({ error: title + " already exists in library" });
-      }
-      try {
+        return res.json("title already exists" );
+      }      
         // create new book
         let newBook = new Book({
           title: title,
           comments: [],
           commentcount: 0,
         });
-        await newBook.save();
+        try {
+        const result = await newBook.save();
         // respond with saved book
-        return res.json({ title: newBook.title, id: newBook._id });
-      } catch {
+        res.json({ title: result.title, _id: result._id });
+      } catch (err) {
         // Handle errors
-        res.status(500).json({ error: "Internal server error" });
+        res.status(500).send({ error: "book POST error" });
       }
     })
 
-    .delete(function (req, res) {
+    .delete(async function (req, res) {
       //if successful response will be 'complete delete successful'
-      console.log("route/api/books", "delete req");
-
+      // delete all books
+      try {
+          const deletedAllBooks = await Book.deleteMany();
+          console.log("deleted all books", deletedAllBooks);          
+            res.send('complete delete successful');          
+      } catch (err) {
+          res.send("err")
+      }
     });
 
   app
@@ -83,48 +95,59 @@ module.exports = function (app) {
     .get( async function (req, res) {
       let bookid = req.params.id;
       try {
+        if( bookid == '' || !bookid) {
+          return res.json("no book exists");
+        }
         const foundBook = await Book.findById(bookid);
         if(!foundBook) {
-          return res.json("No book exists");
-        }
+          return res.json('no book exists');
+        }        
         return res.json( foundBook );
       } catch (error) {
-          res.json("error");
+          res.status(500).json("api/book/:id get error");
       }
       //json res format: {"_id": bookid, "title": book_title, "comments": [comment,comment,...]}
     })
 
     .post(async function (req, res) {
       let bookid = req.params.id;
-      let comment = req.body.comment;
-      //console.log('Comment: ',comment);
-      //console.log(bookid);       
-      try {  
-        if(!comment) {
+      let comment = req.body.comment;       
+      if(!comment || comment == '') {
           return res.json("missing required field comment")
-        }  
+      }
+      try {    
         const foundBook = await Book.findById(bookid);
         if (!foundBook) {
           return res.json("no book exists");          
         }
-        foundBook.comments.push(comment); // add comment
-        foundBook.commentcount += 1;      // increment count
+        // add comment  to array
+        foundBook.comments.push(comment);
+        // increment count 
+        foundBook.commentcount += 1;      
         // save updated book
         const updatedBook = await foundBook.save();
-        return res.json({ updatedBook });
+        res.json({ updatedBook });
       } catch (error) {
-         res.json("error updating book");
+        /// check if format ok in sending ALL fields back
+         res.send("no book exists");
       }      
-      //json res format same as .get
+      //json res format same as .get???
     })
 
     .delete(async function (req, res) {
-      let bookid = req.params.id;      
+      let bookid = req.params.id;
+      console.log("delete book: ", bookid);     
       try {
-        await Book.deleteOne({_id: bookid});
-        res.json('delete successful');
+        // maybe delete the if an just use the catch??
+        if(!bookid || bookid == '') {
+          return res.json('no book exists');
+        }
+        // check the format of the parameter required in deleteOne
+        const result = await Book.deleteOne({_id: bookid});
+        // maybe try : if(result) { res.json('delete successful)} else { res.json('no bokk exists')}
+        res.send('delete successful');
       } catch (error) {
-        res.json('no book found'); 
+        res.send('no book exists'); 
       }
     });
 };
